% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaLookup.R
\name{areaLookup}
\alias{areaLookup}
\title{Calculate cross sectional area for a a range of heights}
\usage{
areaLookup(xs, interval = 0.1, window = c(0, 0))
}
\arguments{
\item{xs}{dataframe, with columns "chain" and "rl".}

\item{interval}{Numeric for the interval between height steps}

\item{window}{vector with a length of 2 with the minimum and maximum values}
}
\value{
dataframe of height and area
}
\description{
Applies calcArea at intervals to produce a lookup table
}
\examples{
xs <- data.frame(chain = c(1,2,3,4,5,6,7),
    rl = c(10,9,7,4,8,9,11)
)
# default
areaLookup(xs)
# specify window
areaLookup(xs, window = c(7,9))
# 1m interval
areaLookup(xs, interval = 1)
}
