% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockList.R
\name{blockList}
\alias{blockList}
\title{Creates a list of blocks, with a start time, end time and duration}
\usage{
blockList(gaps, ts)
}
\arguments{
\item{tsdata}{A dataframe, with first column POSIXct time}

\item{minGap}{Numeric, number of minutes to record a gap}
}
\value{
dataframe listing each gap start, end and duration.
}
\description{
Creates a list of blocks, with a start time, end time and duration
}
\examples{


library(dplyr)
# create a df of hourly times

randomMinutes <- rnorm(1000, mean = 60, sd = 30) # create 1000 random intervals
randomMinutes[randomMinutes < 1] <- 1 # set less than 1 minute to 1
randomMinutes <- as.POSIXct("2021-01-01") + cumsum(randomMinutes)*60 # add to a start point

# Gaplist of a single column dataframe
head(GapList(data.frame(randomMinutes), 60)) # 60 minutes minimum gap
head(GapList(data.frame(randomMinutes), 120)) # 120 minutes minimum gap
head(GapList(data.frame(randomMinutes), 180)) # 180 minutes minimum gap

# build a data frame of times and values
df <- data.frame(Time = randomMinutes,
                data = cumsum(rnorm(1000, mean = 0, sd = 1))   # add a row of random numbers
)

plot(df)
head(GapList(df, 120)) # Gaplist of a multi column dataframe (same result)



}
