% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcArea.R
\name{calcArea}
\alias{calcArea}
\title{Calculate Cross sectional area for a particular height}
\usage{
calcArea(xscalc, ght)
}
\arguments{
\item{xscalc}{dataframe, with columns "chain" and "rl".}

\item{ght}{Numeric}
}
\value{
value
}
\description{
Converts continuous BMI variable to BMI groups.
}
\examples{
xs <- data.frame(chain = c(1,2,3,4,5,6,7),
    rl = c(10,9,7,4,8,9,11)
)
calcArea(xs, 1)
calcArea(xs, 8)
calcArea(xs, 11)
calcArea(xs, 20)

}
