% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapList.R
\name{GapList}
\alias{GapList}
\title{Interpolates for xy coordinates}
\usage{
GapList(tsdata, minGap = 60)
}
\arguments{
\item{tsdata}{A dataframe, with first column POSIXct time}

\item{minGap}{Numeric, number of minutes to record a gap}
}
\value{
dataframe listing each gap start, end and duration.
}
\description{
To interpolate values (xaxis) at a location (yaxis), in a series of data.
}
\examples{


library(dplyr)
# create a df of hourly times

randomMinutes <- rnorm(1000, mean = 60, sd = 30) # create 1000 random intervals
randomMinutes[randomMinutes < 1] <- 1 # set less than 1 minute to 1
randomMinutes <- as.POSIXct("2021-01-01") + cumsum(randomMinutes)*60 # add to a start point

GapList(data.frame(randomMinutes), 60) # Gaplist of a single column dataframe
GapList(data.frame(randomMinutes), 120) # Gaplist of a single column dataframe
GapList(data.frame(randomMinutes), 180) # Gaplist of a single column dataframe

# build a data frame of times and values
df <- data.frame(Time = randomMinutes,
                data = cumsum(rnorm(1000, mean = 0, sd = 1))   # add a row of random numbers
)

plot(df) # Gaplist of a single column dataframe
GapList(df, 120) # Gaplist of a multi column dataframe (same result)



}
