\name{HydroTopmodelParameters-class}
\docType{class}
\alias{HydroTopmodelParameters-class}
\alias{plot,HydroTopmodelParameters,ANY-method}
\alias{print,HydroTopmodelParameters-method}

\title{Class "HydroTopmodelParameters"}
\description{	 Implementation of
\code{\linkS4class{HydroModelParameters}} for Topmodel }
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("HydroTopmodelParameters", ...)}.
 
ToDo
}
\section{Slots}{
   see \code{"\linkS4class{HydroModelParameters}"}
	 }
\section{Extends}{
Class \code{"\linkS4class{HydroModelParameters}"}, directly.
}
\section{Methods}{
  \describe{
    \item{plot}{\code{signature(x = "HydroTopmodelParameters", y = "ANY")}: ... }
    \item{print}{\code{signature(x = "HydroTopmodelParameters")}: ... }
	 }
}
\references{ thePaper }
\author{ Dominik Reusser, Wouter Buytaert }

\seealso{
	\code{\link{topmodel}}
}
\examples{
showClass("HydroTopmodelParameters")
}
\keyword{classes}
