% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpSeries.R
\name{interpSeries}
\alias{interpSeries}
\title{Interpolates for xy coordinates}
\usage{
interpSeries(series, location)
}
\arguments{
\item{series}{A dataframe, with x,y coordinates}

\item{location}{Numeric, location along the line (y-axis) of which to interpolate on the x-axis}
}
\value{
dataframe
}
\description{
To interpolate values (xaxis) at a location (yaxis), in a series of data.
}
\examples{
xs <- data.frame(chain = c(1,2,3,4,5,6,7),
    rl = c(10,9,7,4,8,9,11)
)
# Interpolate the cross section at an RL of 8.5
interpSeries(xs, 8.5)

}
