% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manningsEqn.R
\name{manningsEqn}
\alias{manningsEqn}
\title{Applies Mannings equation across a range of heights}
\usage{
manningsEqn(xssummary, slope = 0.0015, n = 0.05)
}
\arguments{
\item{xssummary}{datamframe, output from xsSummary()}

\item{slope}{numeric, flood slope of river}

\item{n}{numeric, roughness coefficient of river}
}
\value{
dataframe of xs summary with added discharge
}
\description{
Adds discharge to a discharge summary
}
\examples{
xs <- data.frame(chain = c(0,20,30,40,50,60,70),
    rl = c(10,9,7,4,8,9,11)
)
summary <- xsSummary(xs, by = 1)
manningsEqn(summary, n=0.05, s=0.0015)

}
