% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTranslate.R
\name{matrixTranslate}
\alias{matrixTranslate}
\title{Translate x,y,z coordinates by offset}
\usage{
matrixTranslate(m, translation = c(0, 0, 0))
}
\arguments{
\item{m}{dataframe, with columns "x", "y", "z"}

\item{x}{float, x axis offset}

\item{y}{float, x axis offset}

\item{z}{float, x axis offset}
}
\value{
dataframe, translated input dataframe
}
\description{
Translate x,y,z coordinates by offset
}
\examples{
m <- data.frame(x = c(2.5,2,3,1), y = c(3,5,10,15), z = c(3,2,1,3))
# plotting y ( chainage ) against x (alignment)
plot(y ~ x, m, xlim = c( min(m$x, m$y) - max( m$x, m$y ) - 10 , max(m$x, m$y)*2 + 10 ) ,
   ylim =c( min(m$x, m$y) - max( m$x, m$y ) - 10 , max(m$x, m$y)*2 + 10 ))
abline (1,0)
abline (v = 0)
points(y ~ x, matrixTranslate(m, c(10,5,0)), col="red")

}
