\name{mopex}
\docType{data}
\alias{fuse.DATA}
\alias{fuse.parameters}
\title{Hydrological dataset to run FUSE}
\description{
This is an example dataset containing the following data:  
\tabular{ll}{
    DATA       \tab Observed precipitation, potential evapo-transpiration and discharge (mm/day)\cr
    parameters \tab Values for the FUSE parameters\cr
}
The dataset refers to the MOPEX catchment n. 03054500 (Clark et al., 2008).
}
\usage{data}
\format{data frame}
%\source{}
\references{
Clark M. P., SlaterA. G., Rupp D. E., Woods R. A., Vrugt J. A., Gupta H. V., Wagener T. and Hay L. E. (2008), Framework for 	Understanding Structural Errors (FUSE): A modular framework to diagnose differences between hydrological models, Water Resour. Res. 44 p. 91-94 
}
\keyword{datasets}

