% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhasedAreas.R
\name{PhasedAreas}
\alias{PhasedAreas}
\title{Calculate cross sectional area for a a range of periods}
\usage{
PhasedAreas(RatePer, xsdf, stagedf)
}
\arguments{
\item{RatePer}{dataframe, with columns "PeriodStart", "End", "Phased", "TableNum"
PeriodStart (Posixct) start time of the rating period
Phased (Numeric) 1 or 0.
TableNum (Alphanumeric) the section ID number to be applied for this time}

\item{xsdf}{dataframe, with columns "CHAIN", "RL" and "n"
CHAIN (Numeric) cross section chainages
RL (Numeric) cross section Reduced levels
n (Aphanumeric) a unique cross section ID number for a particuar cross section.
n is same as TableNum in RatePer}

\item{stagedf}{dataframe, with columns for a timestamp (posixct) and a height (Numeric)}
}
\value{
Vector of areas, from a lookup table at each Stage timestamp,
therefore will always merge with Stage.
}
\description{
Applies a rating table to height data for a time period
If the phased box is set to true, the result is based on a time factor between the
current rating table and the next rating table.
}
\examples{
library(ggplot2)
# Set cross sections
xs <- data.frame(CHAIN = c(1,2,3,4,5,6,7),
    RL = c(10,9,7,4,8,9,11)
)
xs2 <- data.frame(CHAIN = c(1,2,3,4,5,6,7),
    RL = c(10,9,4,4,4,9,11)
)
xs$n <- "xs1"
xs2$n <- "xs2"
XSDF <- rbind(xs, xs2)
# plot each cross section
ggplot(XSDF, aes(CHAIN, RL, colour=n))+
 geom_line()+
 geom_point()

# Set rating periods
Periods <- data.frame(
    PeriodStart = as.POSIXct(c("2020-01-01 00:00", "2021-01-01 00:00", "2022-01-01 00:00")),
    Phased = c(1,0,1),
    TableNum = c("xs1", "xs2", "xs1")
    )

# Create a constant stage over time
Stage <- data.frame(
    timestamp = seq(as.POSIXct("2020-01-01 00:00"), as.POSIXct("2024-01-01 00:00"),
        by = 60*60*24), # daily timestamps
    height = 7
  )

# Apply phased rating tables to stage
Stage$Area <- PhasedAreas(Periods, XSDF, Stage)
# Heights covering entire rating periods
plot(Stage$timestamp, Stage$Area)

# Subsection Stage
Stage <- data.frame(
    timestamp = seq(as.POSIXct("2020-06-01 00:00"), as.POSIXct("2024-01-01 00:00"),
        by = 60*60*24), # daily timestamps
    height = 7
  )

# Apply phased rating tables to stage
Stage$Area <- PhasedAreas(Periods, XSDF, Stage)
# Plots starting after the first rating period has started
plot(Stage$timestamp, Stage$Area)
# Subsection Stage
Stage <- data.frame(
    timestamp = seq(as.POSIXct("2019-01-01 00:00"), as.POSIXct("2024-01-01 00:00"),
        by = 60*60*24), # daily timestamps
    height = 7
  )

# Apply phased rating tables to stage
Stage$Area <- PhasedAreas(Periods, XSDF, Stage)
# Heights starting before rating periods - No values before period starts.
plot(Stage$timestamp, Stage$Area)

# Set rating periods to always phased change
Periods <- data.frame(
    PeriodStart = as.POSIXct(c("2020-01-01 00:00", "2021-01-01 00:00",
        "2022-01-01 00:00", "2024-01-01 00:00")),
    Phased = c(1,1,1,1),
    TableNum = c("xs1", "xs2", "xs1", "xs2")
    )

# Apply phased rating tables to stage
Stage$Area <- PhasedAreas(Periods, XSDF, Stage)
# Heights covering entire rating periods
plot(Stage$timestamp, Stage$Area)

}
