% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processEMD.R
\name{processEMD}
\alias{processEMD}
\title{Processes EMD blocks at at time, and then interpolates between}
\usage{
processEMD(tsdata, tsblocks = 0, method = "ceemdan")
}
\arguments{
\item{tsdata}{dataframe, with posixct for time and values, and a defined interval i.e. hourly}

\item{tsblocks}{tsblocks, data.frame(Start = posixct, End=posixct)}

\item{method}{"emd" or "ceemdan"}
}
\value{
dataframe of filtered/decomposed data
}
\description{
Uses predefined ceemdan parameters, specifically for estuaries
}
\examples{

library(ccInterp)
library(ggplot2)

# make some random data
randomMinutes <- rnorm(10000, mean = 60, sd = 100) # create 1000 random intervals
randomMinutes <- randomMinutes[randomMinutes > 30]  # cull short or negative intervals
randomMinutes <- as.POSIXct(cumsum(randomMinutes)*60, origin="2021-01-01")
# build a data frame of times and values
df <- data.frame(Time = randomMinutes,
               # add a row of random numbers
               data = cumsum(rnorm(length(randomMinutes), mean = 0, sd = 10))
)

# change to hourly
hourlydf <- changeInterval(df, Interval = "Hourly", option="inst", offset = 30)

gaps <- GapList(df, 360) # 6 hour gaps
blocks <- blockList(gaps, df)
data <- processEMD(hourlydf, blocks)
plot(df, col="grey")
lines(data$time, data$a)
lines(data$time, data$b, col="blue")
lines(data$time, data$c, col="red")
}
