% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhydro.R
\docType{package}
\name{rhydro}
\alias{rhydro-package}
\alias{rhydro}
\title{Tools for hydrographic data processing in R}
\description{
Hydro tools
}
\details{
Such as
And
Things
}
\author{
Stephen Wallace \email{stephen.wallace@des.qld.gov.au}
}
