% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBlocks.R
\name{splitBlocks}
\alias{splitBlocks}
\title{Omit or include data between dates}
\usage{
splitBlocks(blocklist, newdata, typekey = NULL, include = 1)
}
\arguments{
\item{blocklist}{output from blockList/GapList}

\item{newdata}{dataframe, timestamp, and one or more columns of data}

\item{typekey}{optional, for interpolating boundaries
a vector descring the interpolation of the
respective columns.
0 = linear
1 = maximum neighbor
NULL = linear for all
Where not NULL, must be length of input dataframes}

\item{include}{to either include the 'newdata' or exclude it
based on blocklist start/end.
i.e. when include is 1, the 'newdata' is chunked to only dates within
the 'blocklist', including interpolated block boundaries
when include is 0, 'newdata' is chunked to only dates outside the blocklist.}
}
\value{
dataframe of the original data within, or outside of blocklist.
}
\description{
Used for merging datasets, or deleting data during gaps
}
\examples{
library(dplyr)


ts <- ccInterp::StevesCoolRandomTS(smoothed = FALSE, obs = 10000 )
plot(ts$Time, ts$Noise)
gaps <- rhydro::GapList(ts, minGap = 120)

# include between gaps
splitTS <- splitBlocks(gaps, ts, include = 1)
plot(splitTS$Time, splitTS$Noise, col="red")
#exclude between gaps
splitTS <- splitBlocks(gaps, ts, include = 0)
points(splitTS$Time, splitTS$Noise, col="blue")

splitTS <- blockList(gaps, ts) \%>\% # convert gaps to blocks of ts
splitBlocks(ts, include = 1) # inclide blocks only
plot(splitTS$Time, splitTS$Noise, col="blue")

splitTS <- blockList(gaps, ts) \%>\% # convert gaps to blocks of ts
splitBlocks(ts, include = 0)
points(splitTS$Time, splitTS$Noise, col="red")
}
