% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsAlignment.R
\name{xsAlignment}
\alias{xsAlignment}
\title{Rotate x,y,z coordinates around an axis}
\usage{
xsAlignment(xs, rp = NULL, originRP, alignRP, angle = NULL, origin = NULL)
}
\arguments{
\item{xs}{dataframe, columns required x, y, z, rpid (+optional meta)
Where multiple lines for the same rpid exist i.e. closing a cross section, it gets the mean}

\item{rp}{data.frame, columns required x, y, z, rpid (+optional meta)}

\item{originRP}{string, name of primary reference point (that new xs is shifted to)}

\item{alignRP}{string, name of secondary reference point ( that determiens alignment angle )}

\item{angle}{float, override angle}

\item{origin}{vector, length 3, override origin, 1=x, 2=y, 3=z}
}
\value{
dataframe, rotated
}
\description{
Rotate x,y,z coordinates around an axis
}
\examples{

# NOT RUN #
#DRL <- getXSSurvey(client, "1080025", bucket = "tsdata3")
#
data(DRLXS) # example site

head(DRL$RP)
head(DRL$points[[2]])

# align DRL points to RP's RP1 and RP2 for a specific cross section
aligned <- xsAlignment(DRL$points[[2]],
                       DRL$RP,
                       originRP = "RP1",
                       alignRP = "RP2")

head(aligned)

# align all cross sections
aligned <- lapply(DRL$points, xsAlignment,
                  rp = DRL$RP,
                  originRP = "RP1",
                  alignRP = "RP2")

}
