% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsSummary.R
\name{xsSummary}
\alias{xsSummary}
\title{Calculate cross section stats for a range of heights}
\usage{
xsSummary(xs, from = -1, to = -1, by = 0.1)
}
\arguments{
\item{xs}{dataframe, with columns "chain" and "rl".}

\item{from}{numeric, start gauge height range}

\item{to}{numeric, end gauging height range}

\item{by}{numeric, step increment for summary range}
}
\value{
dataframe of gauge heights and cross section stats
}
\description{
Area, wetted perimeter, hydraulic radius
}
\examples{
xs <- data.frame(chain = c(1,2,3,4,5,6,7),
    rl = c(10,9,7,4,8,9,11)
)
xsSummary(xs)
xsSummary(xs, by = 1)
xsSummary(xs, from = 6, to = 9, by = 0.05)

}
