% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockList.R
\name{blockList}
\alias{blockList}
\title{Creates a list of blocks, with a start time, end time and duration}
\usage{
blockList(gaps, ts, minblock = 360)
}
\arguments{
\item{gaps}{A dataframe, with first column POSIXct time}

\item{ts}{Numeric, number of minutes to record a gap}

\item{minblock}{Numeric, minimum length of a "block" in minutes}
}
\value{
dataframe listing each gap start, end and duration.
}
\description{
Creates a list of blocks, with a start time, end time and duration
}
\examples{

# create a df of hourly times

randomMinutes <- rnorm(1000, mean = 60, sd = 30) # create 1000 random intervals
randomMinutes[randomMinutes < 1] <- 1 # set less than 1 minute to 1
randomMinutes <- as.POSIXct("2021-01-01") + cumsum(randomMinutes)*60 # add to a start point

# Gaplist of a single column dataframe
gaps <- GapList(data.frame(randomMinutes), 120) # 180 minutes minimum gap

# build a data frame of times and values
df <- data.frame(Time = randomMinutes,
                data = cumsum(rnorm(1000, mean = 0, sd = 1))   # add a row of random numbers
)


ten <- blockList(gaps, df, minblock = 10) # at least 10 minutes per block
day <- blockList(gaps, df, minblock = 1440) # day long blocks only

head(ten)
head(day)

}
