% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logInterpolate.R
\name{logInterpolate}
\alias{logInterpolate}
\title{Lograrithmic interpolation between points}
\usage{
logInterpolate(rating, num, logOffset = 0, base = 10)
}
\arguments{
\item{rating}{dataframe, with columns for reference and a lookup value,
i.e. a height reference, and a lookup area.}

\item{num}{Numeric or vector for desired values}

\item{logOffset}{Numeric to specify the log offset used.
Leave as zero to set as the maximum value that the result is zero,
i.e. one row less than the first positive value}
}
\value{
dataframe of height and area
}
\description{
Intpolates a set of points logarithmically
    Either sets its own log offset based on inputs, or as determined
}
\examples{
rating <- data.frame(level = c(2,2.5,4,6,10),
  discharge = c(0.1,10,100,1000,4000)
                      )

logInterpolate(rating, 3.5)
logInterpolate(rating, c(3.5, 5.5, 9.5))

plot(rating)
values <- pretty(range(rating$level), n=20)
lines(values,
 logInterpolate(rating,values),
 col="red"
)

}
