% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logInterpolate.R
\name{logInterpolate}
\alias{logInterpolate}
\title{Calculate cross sectional area for a a range of heights}
\usage{
logInterpolate(rating, num, logOffset = 0)
}
\arguments{
\item{rating}{dataframe, with columns for height and area}

\item{num}{Numeric or vector for desired values}

\item{logOffset}{Numeric to specify the log offset used.
Leave as zero to set as the maximum value that the result is zero,
i.e. one row less than the first positive value}
}
\value{
dataframe of height and area
}
\description{
Applies calcArea at intervals to produce a lookup table
    Below the lowest postive number in the second column is
    set to zero.  It is not interpolated to zero.
}
\examples{
xs <- data.frame(chain = c(1,2,3,4,5,6,7),
    rl = c(10,9,7,4,8,9,11)
)
lookupTable <- areaLookup(xs, interval = 1)
# with default used for logOffset ( 4 in this case)
logInterpolate(lookupTable, 7.5)
# plot cross section
plot(xs)
# plot 1m interval lookup table
plot(lookupTable)
# plot log interpolated values
heights <- seq(min(lookupTable$height), max(lookupTable$height), by = 0.01 )
lines(heights, logInterpolate(lookupTable, heights), col="red")

}
